update selector line frame based on selectedIndex
There was an issue with creating this using a frame of zero and then
using AutoLayout to constraint it, and setting the selected index to
anything aside from the first one. The line would not draw, because
selector.frame.origin in layoutSubviews was still 0.

This changes uses the index of the selected item to draw the selector
line.