remove appearance vars that will be unused
This is going to switch to the titles being left aligned by default,
and thus these 2 properties will not be used. If/when we change this to
support center aligned titles, then we’ll introduce properties to
control that in a more sophisticated way.

This also adds a button to reset the view in the Demo.